+++ lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:qr_scanner/screens/authenticate/signup.dart';
-import 'package:qr_scanner/screens/authenticate/userlogin.dart';
 import 'package:qr_scanner/screens/wrapper.dart';
 import 'package:qr_scanner/services/auth.dart';
 import 'models/user.dart';
@@ -10,7 +9,7 @@ import 'screens/home/history.dart';
 import 'screens/authenticate/login.dart';
 import 'package:provider/provider.dart';
 
-void main() async {
+Future main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
   runApp(
@@ -19,13 +18,12 @@ void main() async {
       initialData: null,
       child: MaterialApp(
         debugShowCheckedModeBanner: false,
-        home: Wrapper(),
+        home: const Wrapper(),
         routes: {
           '/home': (context) => HomePage(),
           '/history': (context) => HistoryPage(),
           '/signup': (context) => SignUp(),
           '/login': (context) => LoginPage(),
-          '/userlogin': (context) => login(),
         },
       ),
     ),
