+++ lib/screens/authenticate/login.dart
@@ -1,6 +1,8 @@
 // login screen
 
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:qr_scanner/services/auth.dart';
 
 class LoginPage extends StatefulWidget {
@@ -13,88 +15,303 @@ class LoginPage extends StatefulWidget {
 class _LoginPageState extends State<LoginPage> {
   final AuthService _auth = AuthService();
 
+  Widget _buildEmailTF() {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: <Widget>[
+        const Text(
+          'Email ID',
+          style: TextStyle(
+            fontFamily: 'Product Sans',
+            color: Color(0xFFA0A0A0),
+          ),
+        ),
+        SizedBox(height: 10.0),
+        Container(
+          alignment: Alignment.centerLeft,
+          decoration: BoxDecoration(
+            color: Colors.white,
+            borderRadius: BorderRadius.circular(10.0),
+            boxShadow: const [
+              BoxShadow(
+                color: Colors.black12,
+                blurRadius: 6.0,
+                offset: Offset(0, 2),
+              ),
+            ],
+          ),
+          height: 60.0,
+          child: const TextField(
+            keyboardType: TextInputType.emailAddress,
+            style: TextStyle(
+              color: Colors.white,
+              fontFamily: 'Product Sans',
+            ),
+            decoration: InputDecoration(
+              border: InputBorder.none,
+              contentPadding: EdgeInsets.only(top: 14.0),
+              prefixIcon: Icon(
+                Icons.email,
+                color: Color(0XFFFF7D54),
+              ),
+              hintText: 'Enter your Email',
+              hintStyle: TextStyle(
+                fontFamily: 'Product Sans',
+                color: Color(0xFFA0A0A0),
+              ),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildPasswordTF() {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: <Widget>[
+        const Text(
+          'Password',
+          style: TextStyle(
+            fontFamily: 'Product Sans',
+            color: Color(0xFFA0A0A0),
+          ),
+        ),
+        SizedBox(height: 10.0),
+        Container(
+          alignment: Alignment.centerLeft,
+          decoration: BoxDecoration(
+            color: Colors.white,
+            borderRadius: BorderRadius.circular(10.0),
+            boxShadow: const [
+              BoxShadow(
+                color: Colors.black12,
+                blurRadius: 6.0,
+                offset: Offset(0, 2),
+              ),
+            ],
+          ),
+          height: 60.0,
+          child: const TextField(
+            obscureText: true,
+            style: TextStyle(
+              color: Colors.white,
+              fontFamily: 'Product Sans',
+            ),
+            decoration: InputDecoration(
+              border: InputBorder.none,
+              contentPadding: EdgeInsets.only(top: 14.0),
+              prefixIcon: Icon(
+                Icons.lock,
+                color: Color(0XFFFF7D54),
+              ),
+              hintText: 'Enter your Password',
+              hintStyle: TextStyle(
+                fontFamily: 'Product Sans',
+                color: Color(0xFFA0A0A0),
+              ),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildForgotPasswordBtn() {
+    return Container(
+      alignment: Alignment.centerRight,
+      child: FlatButton(
+        onPressed: () => print('Forgot Password Button Pressed'),
+        padding: EdgeInsets.only(right: 0.0),
+        child: const Text(
+          'Forgot Password?',
+          style: TextStyle(
+            color: Colors.grey,
+            fontWeight: FontWeight.bold,
+            fontFamily: 'Product Sans',
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildLoginBtn() {
+    return Container(
+      padding: const EdgeInsets.symmetric(vertical: 10.0),
+      width: double.infinity,
+      child: RaisedButton(
+        elevation: 3.0,
+        onPressed: () {
+          print('Login Button Pressed');
+          Navigator.of(context).pushNamed('/home');
+        },
+        padding: const EdgeInsets.all(15.0),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(30.0),
+        ),
+        color: const Color(0XFFFF7D54),
+        child: const Text(
+          'LOGIN',
+          style: TextStyle(
+            color: Colors.white,
+            letterSpacing: 1.5,
+            fontSize: 18.0,
+            fontWeight: FontWeight.bold,
+            fontFamily: 'Product Sans',
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildSignInWithText() {
+    return const Padding(
+      padding: EdgeInsets.all(20.0),
+      child: Text(
+        '- OR -',
+        style: TextStyle(
+          color: Colors.grey,
+          fontWeight: FontWeight.w400,
+        ),
+      ),
+    );
+  }
+
+  Widget _buildSocialBtnRow() {
+    return Container(
+      width: 260.0,
+      child: ElevatedButton.icon(
+        onPressed: () {},
+        style: ElevatedButton.styleFrom(
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(30.0),
+          ),
+          padding: EdgeInsets.all(10.0),
+          primary: Colors.grey[600],
+        ),
+        icon: const FaIcon(
+          FontAwesomeIcons.google,
+        ),
+        label: Text('Sign In with Google'),
+      ),
+    );
+  }
+
+  Widget _buildSignupBtn() {
+    return Padding(
+      padding: const EdgeInsets.all(20.0),
+      child: GestureDetector(
+        onTap: () {
+          print('Sign Up Button Pressed');
+          Navigator.of(context).pushNamed('/signup');
+        },
+        child: RichText(
+          text: const TextSpan(
+            children: [
+              TextSpan(
+                text: 'Don\'t have an Account? ',
+                style: TextStyle(
+                  fontFamily: 'Product Sans',
+                  color: Colors.grey,
+                  fontSize: 18.0,
+                  fontWeight: FontWeight.w400,
+                ),
+              ),
+              TextSpan(
+                text: 'Sign Up',
+                style: TextStyle(
+                  fontFamily: 'Product Sans',
+                  color: Colors.black,
+                  fontSize: 18.0,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        elevation: 0.0,
-        leading: IconButton(
-            onPressed: () => Navigator.of(context).pop(),
-            color: Colors.black,
-            iconSize: 35.0,
-            icon: const Icon(Icons.arrow_back_rounded)),
-      ),
-      body: Center(
-        child: Column(
-          children: [
-            Container(
-              padding: const EdgeInsets.fromLTRB(50.0, 0.0, 50.0, 0.0),
-              child: Image.asset(
-                'assets/logo.png',
-                height: 170,
-                width: 170,
-              ),
-            ),
-            Container(
-              padding: const EdgeInsets.all(5.0),
-              child: const Text(
-                'Welcome Back!',
-                style: TextStyle(
-                    fontFamily: 'Product Sans',
-                    fontWeight: FontWeight.bold,
-                    fontSize: 26.0,
-                    color: Color(0xFF1E1E1E)),
-              ),
-            ),
-            Container(
-              padding: const EdgeInsets.fromLTRB(50.0, 10.0, 50.0, 20.0),
-              child: const Text(
-                //Need to change this line
-                'Log in to your existent account of QR Scanner',
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  fontFamily: 'Product Sans',
-                  height: 1.4,
-                  color: Color(0xFFA0A0A0),
+      body: AnnotatedRegion<SystemUiOverlayStyle>(
+        value: SystemUiOverlayStyle.light,
+        child: GestureDetector(
+          onTap: () => FocusScope.of(context).unfocus(),
+          child: Stack(
+            children: <Widget>[
+              Container(
+                height: double.infinity,
+                child: SingleChildScrollView(
+                  physics: const AlwaysScrollableScrollPhysics(),
+                  padding: const EdgeInsets.fromLTRB(40.0, 30.0, 40.0, 30.0),
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      Container(
+                        width: double.infinity,
+                        height: 60.0,
+                        child: Align(
+                          alignment: Alignment.topRight,
+                          child: FlatButton(
+                            child: const Text(
+                              'Guest Mode',
+                              style: TextStyle(
+                                  fontFamily: 'Product Sans',
+                                  fontWeight: FontWeight.bold,
+                                  fontSize: 15.0,
+                                  color: Color(0xFF1E1E1E)),
+                            ),
+                            onPressed: () async {
+                              dynamic result = await _auth.signInAnon();
+                              if (result == null) {
+                                print('Error signing in');
+                              } else {
+                                print('Signed In');
+                                print(result.uid);
+                              }
+                            },
+                          ),
+                        ),
+                      ),
+                      Container(
+                        decoration: BoxDecoration(
+                          color: Colors.grey[100],
+                          borderRadius: BorderRadius.circular(25.0),
+                        ),
+                        padding: const EdgeInsets.all(30.0),
+                        child: Column(
+                          children: [
+                            const Text(
+                              'Sign In',
+                              style: TextStyle(
+                                color: Colors.black,
+                                fontFamily: 'Product Sans',
+                                fontWeight: FontWeight.bold,
+                                fontSize: 30.0,
+                              ),
+                            ),
+                            const SizedBox(height: 30.0),
+                            _buildEmailTF(),
+                            const SizedBox(height: 30.0),
+                            _buildPasswordTF(),
+                            _buildForgotPasswordBtn(),
+                            _buildLoginBtn(),
+                          ],
+                        ),
+                      ),
+                      _buildSignInWithText(),
+                      _buildSocialBtnRow(),
+                      _buildSignupBtn(),
+                    ],
+                  ),
                 ),
-              ),
-            ),
-            Container(
-              padding:
-                  const EdgeInsets.symmetric(vertical: 20.0, horizontal: 50.0),
-              child: RaisedButton(
-                child: const Text('Guest Login'),
-                onPressed: () async {
-                  dynamic result = await _auth.signInAnon();
-                  if (result == null) {
-                    print('Error signing in');
-                  } else {
-                    print('Signed In');
-                    print(result.uid);
-                  }
-                },
-              ),
-            ),
-            Container(
-              padding:
-                  const EdgeInsets.symmetric(vertical: 20.0, horizontal: 50.0),
-              child: RaisedButton(
-                child: const Text('User Login'),
-                onPressed: () => Navigator.pushNamed(context, '/userlogin'),
-//                  async {
-//                    dynamic result = await _auth.signInAnon();
-//                    if (result == null) {
-//                      print('Error signing in');
-//                    } else {
-//                      print('Signed In');
-//                      print(result.uid);
-//                    }
-//                  }),
-              ),
-            ),
-          ],
+              )
+            ],
+          ),
         ),
       ),
     );
