+++ pubspec.yaml
@@ -36,9 +36,17 @@ dependencies:
   cupertino_icons: ^1.0.2
   barcode_scan2: ^4.2.0
   url_launcher: ^6.0.18
+
+  # Firebase Sign In
   firebase_auth: ^3.3.6
   firebase_core : ^1.10.2
   cloud_firestore: ^3.1.7
+  google_sign_in: ^5.2.3
+
+  # Google Icon
+  font_awesome_flutter: ^9.2.0
+
+  # State Management
   provider: ^6.0.2
 
 dev_dependencies:
